simplify, regen types. remove legacy docs
