+++ modules/async-icq/testing/demo-simapp/.gitignore
@@ -4,3 +4,4 @@ release/
 .idea/
 .vscode/
 .DS_Store
+github.com
