+++ modules/async-icq/testing/demo-simapp/buf.gen.yaml
@@ -1,0 +1,8 @@
+version: v1
+plugins:
+  - name: gocosmos
+    out: .
+    opt: plugins=interfacetype+grpc,Mgoogle/protobuf/any.proto=github.com/cosmos/cosmos-sdk/codec/types
+  - name: grpc-gateway
+    out: .
+    opt: logtostderr=true,allow_colon_final_segments=true
