+++ modules/async-icq/testing/demo-simapp/buf.work.yaml
@@ -1,0 +1,3 @@
+version: v1
+directories:
+  - proto
