--- modules/async-icq/testing/demo-simapp/docs/docs.go
@@ -1,6 +1,0 @@
-package docs
-
-import "embed"
-
-//go:embed static
-var Docs embed.FS
