+++ modules/async-icq/testing/demo-simapp/x/interquery/types/query.pb.gw.go
@@ -224,9 +224,9 @@ func RegisterQueryHandlerClient(ctx context.Context, mux *runtime.ServeMux, clie
 }
 
 var (
-	pattern_Query_Params_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3}, []string{"quasarfinance", "interchainquerydemo", "interquery", "params"}, "", runtime.AssumeColonVerbOpt(true)))
+	pattern_Query_Params_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3}, []string{"quasarfinance", "interchainquerydemo", "interquery", "params"}, "", runtime.AssumeColonVerbOpt(false)))
 
-	pattern_Query_QueryState_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3}, []string{"quasar-finance", "interchainquerydemo", "interquery", "query_state"}, "", runtime.AssumeColonVerbOpt(true)))
+	pattern_Query_QueryState_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3}, []string{"quasar-finance", "interchainquerydemo", "interquery", "query_state"}, "", runtime.AssumeColonVerbOpt(false)))
 )
 
 var (
