+++ modules/async-icq/testing/simapp/cmd/icqd/root.go
@@ -6,7 +6,6 @@ import (
 	"os"
 
 	app "github.com/cosmos/ibc-apps/modules/async-icq/v7/testing/simapp"
-	appparams "github.com/cosmos/ibc-apps/modules/async-icq/v7/testing/simapp/params"
 	"github.com/spf13/cast"
 	"github.com/spf13/cobra"
 
@@ -38,7 +37,7 @@ import (
 
 // NewRootCmd creates a new root command for simd. It is called once in the
 // main function.
-func NewRootCmd() (*cobra.Command, appparams.EncodingConfig) {
+func NewRootCmd() (*cobra.Command, app.EncodingConfig) {
 	encodingConfig := app.MakeEncodingConfig()
 
 	cfg := sdk.GetConfig()
@@ -100,7 +99,7 @@ func initTendermintConfig() *tmcfg.Config {
 	return cfg
 }
 
-func initRootCmd(rootCmd *cobra.Command, encodingConfig appparams.EncodingConfig) {
+func initRootCmd(rootCmd *cobra.Command, encodingConfig app.EncodingConfig) {
 	ac := appCreator{
 		encCfg: encodingConfig,
 	}
@@ -130,7 +129,7 @@ func addModuleInitFlags(startCmd *cobra.Command) {
 }
 
 // genesisCommand builds genesis-related `simd genesis` command. Users may provide application specific commands as a parameter
-func genesisCommand(encodingConfig appparams.EncodingConfig, cmds ...*cobra.Command) *cobra.Command {
+func genesisCommand(encodingConfig app.EncodingConfig, cmds ...*cobra.Command) *cobra.Command {
 	cmd := genutilcli.GenesisCoreCommand(encodingConfig.TxConfig, app.ModuleBasics, app.DefaultNodeHome)
 
 	for _, subCmd := range cmds {
@@ -192,7 +191,7 @@ func txCommand() *cobra.Command {
 }
 
 type appCreator struct {
-	encCfg appparams.EncodingConfig
+	encCfg app.EncodingConfig
 }
 
 func (ac appCreator) newApp(
